+++ src/SharpDeck/Connectivity/IStreamDeckConnectionController.cs
@@ -1,8 +1,7 @@
-﻿namespace SharpDeck
+﻿namespace SharpDeck.Connectivity
 {
     using System.Threading;
     using System.Threading.Tasks;
-    using SharpDeck.Events.Received;
 
     /// <summary>
     /// Provides a connection to a Stream Deck, with additional functionality for managing the connection.
@@ -12,8 +11,7 @@
         /// <summary>
         /// Initiates a connection to the Stream Deck asynchronously.
         /// </summary>
-        /// <param name="registrationParameters">The registration parameters.</param>
         /// <param name="cancellationToken">The optional cancellation token.</param>
-        Task ConnectAsync(RegistrationParameters registrationParameters, CancellationToken cancellationToken = default);
+        Task ConnectAsync(CancellationToken cancellationToken = default);
     }
 }