+++ src/SharpDeck/Extensions/ServiceCollectionExtensions.cs
@@ -3,7 +3,10 @@
     using System;
     using Microsoft.Extensions.DependencyInjection;
     using Microsoft.Extensions.Logging;
+    using SharpDeck.Connectivity;
     using SharpDeck.Connectivity.Net;
+    using SharpDeck.DependencyInjection;
+    using SharpDeck.Events.Received;
 
     /// <summary>
     /// Provides extension methods for <see cref="IServiceCollection"/>.
@@ -18,14 +21,25 @@
         /// <returns>The modified service collection.</returns>
         public static IServiceCollection AddStreamDeckPlugin(this IServiceCollection services, Action<IStreamDeckPlugin> configure = null)
         {
-            var connectionController = new StreamDeckWebSocketConnection();
-
             return services
-                .AddSingleton<IStreamDeckConnection>(connectionController)
+
+                // Connection with Stream Deck.
+                .AddSingleton(new RegistrationParameters(Environment.GetCommandLineArgs()))
+                .AddSingleton<IStreamDeckConnectionController, StreamDeckWebSocketConnection>()
+                .AddSingleton<IStreamDeckConnection>(serviceProvider => serviceProvider.GetRequiredService<IStreamDeckConnectionController>())
+
+                // Action management, and drill-down.
+                .AddSingleton<IFactory<StreamDeckAction>, StreamDeckButtonFactory<StreamDeckAction>>(serviceProvider => new StreamDeckButtonFactory<StreamDeckAction>(serviceProvider))
+                .AddSingleton<IStreamDeckActionManager, StreamDeckActionManager>()
+
+                // Plug-in.
                 .AddSingleton<IStreamDeckPlugin>(serviceProvider =>
                 {
                     // Construct the plugin, and configure it.
-                    var plugin = new StreamDeckPlugin(connectionController, serviceProvider, serviceProvider.GetService<ILogger<StreamDeckPlugin>>());
+                    var plugin = new StreamDeckPlugin(
+                        serviceProvider.GetRequiredService<IStreamDeckConnectionController>(),
+                        serviceProvider.GetRequiredService<IStreamDeckActionManager>());
+
                     configure?.Invoke(plugin);
 
                     // Initialize the singleton instance.